Make `solve` methods return solution/fail (reference) objects
This replaces the current functionality that forces the user to use a
callback to query the solution or find which assumptions contributed
to failure. The usage of (reference) objects generally makes the system
easier to use and the user doesn't have to care as much about moving
objects.

This also changes the IPASIR-UP implementation to exists through
separate objects. This solves the problem with dynamic typing.