Relaxed `Sized` constraints for encoders
This allows the more liberal usage of non-sized and `dyn` objects of
ClauseDatabase. We introduce the `AsDynClauseDatabase` for placed where
we need to be able to create a `dyn ClauseDatabase` references. This
trait is automatically implemented for all `Sized` types and `dyn`
references, but cannot be automatically implemented for `?Sized` types.