feat: redesign the way that IPASIR solvers are defined in pindakaas
The new design leverages (const) generics to specify the storage
requirements that match the capabilities of a IPASIR solver. For each
solver, traits listing the IPASIR methods are implemented, which will
implement the matching solver trait if the right storage capability is
present.

This design also changes the way `Propagator`s are connected to a
solver. Instead of giving ownership to the solver, a shared reference
is used, allowing the use of the `Propagator` for other (future)
callback functionality.