refactor(pindakaas): remove iset dependency
To remove the dependency, the domain representation in IntVarOrd has
been changed to use a `RangeList` for the domain itself and a `Vec` to
store the literals they represent.

The `bdd` function that tracked a node mapping using an interval map
has been changed to use a sorted vector and binary search.

For convenience in making the changes, the `geq` and `leq` method have
been changed to take only a single integer value and return a `Formula`
object that can then be further combined and encoded.